refactor: Relocate UI components and mappers to feature packages
* Move feature-specific views and tabs from the main iOS app target into dedicated Swift packages (Discover, Library, MyShows, Search, etc.)
* Decentralize data mapping logic by moving `TvManiac` model extensions from `TvManiacKit` to their respective feature modules (Calendar, Library, Profile, Search, and Settings)
* Update access modifiers to `public` for relocated UI components to support cross-module visibility
* Refactor `TvManiacKit` to remove circular dependencies on feature packages
* Update `Package.swift` manifests across the project to reflect the new modular dependency structure
* Relocate supporting UI components like `SortOptionsSheet` and `MyShowsSortOptionsSheet` to their corresponding feature packages
