+++ ios/Packages/Calendar/Sources/Calendar/Mapper+Extensions.swift
@@ -1,0 +1,32 @@
+import TvManiac
+
+// MARK: - Calendar Mapping
+
+public extension TvManiac.CalendarDateGroup {
+    func toSwift() -> SwiftCalendarDateGroup {
+        .init(
+            dateLabel: dateLabel,
+            episodes: episodes.compactMap { $0.toSwift() }
+        )
+    }
+}
+
+public extension TvManiac.CalendarEpisodeItem {
+    func toSwift() -> SwiftCalendarEpisodeItem {
+        .init(
+            showTraktId: showTraktId,
+            episodeTraktId: episodeTraktId,
+            showTitle: showTitle,
+            posterUrl: posterUrl,
+            episodeInfo: episodeInfo,
+            airTime: airTime,
+            network: network,
+            additionalEpisodesCount: additionalEpisodesCount,
+            overview: overview,
+            rating: rating?.doubleValue,
+            votes: votes?.int32Value,
+            runtime: runtime?.int32Value,
+            formattedAirDate: formattedAirDate
+        )
+    }
+}
