+++ ios/Packages/Profile/Sources/Profile/Mapper+Extensions.swift
@@ -1,0 +1,16 @@
+import TvManiac
+
+// MARK: - Trakt List Mapping
+
+public extension TvManiac.TraktListModel {
+    func toSwift() -> SwiftTraktListItem {
+        .init(
+            listId: id,
+            slug: slug,
+            name: name,
+            description: description_,
+            showCountText: showCountText,
+            isShowInList: isShowInList
+        )
+    }
+}
