+++ ios/Packages/Search/Sources/Search/Mapper+Extensions.swift
@@ -1,0 +1,18 @@
+import TvManiac
+
+// MARK: - Search Mapping
+
+public extension TvManiac.ShowItem {
+    func toSwift() -> SwiftSearchShow {
+        .init(
+            tmdbId: tmdbId,
+            traktId: traktId,
+            title: title,
+            overview: overview,
+            status: status,
+            imageUrl: posterImageUrl,
+            year: year,
+            voteAverage: voteAverage?.doubleValue
+        )
+    }
+}
