+++ ios/Packages/TvManiacKit/Package.swift
@@ -19,11 +19,6 @@ let package = Package(
         .package(name: "DesignSystem", path: "../DesignSystem"),
         .package(name: "Models", path: "../Models"),
         .package(name: "Components", path: "../Components"),
-        .package(name: "Calendar", path: "../Calendar"),
-        .package(name: "Search", path: "../Search"),
-        .package(name: "Library", path: "../Library"),
-        .package(name: "Profile", path: "../Profile"),
-        .package(name: "Settings", path: "../Settings"),
         .package(name: "TraktAuthKit", path: "../TraktAuthKit"),
         .package(url: "https://github.com/firebase/firebase-ios-sdk", exact: "12.14.0"),
     ],
@@ -35,11 +30,6 @@ let package = Package(
                 "DesignSystem",
                 "Models",
                 "Components",
-                "Calendar",
-                "Search",
-                "Library",
-                "Profile",
-                "Settings",
                 "TraktAuthKit",
                 .product(name: "FirebaseCore", package: "firebase-ios-sdk"),
             ]
